my_site/requirements.txt
aiohttp==3.6.2
aioredis==1.3.1
amqp==2.6.1
ansicon==1.89.0
anyjson==0.3.3
arrow==0.16.0
asgiref==3.3.0
async-timeout==3.0.1
attrs==20.2.0
autobahn==20.7.1
Automat==20.2.0
autopep8==1.5.4
beautifulsoup4==4.9.1
billiard==3.6.3.0
blessed==1.17.10
cachetools==4.1.1
celery==4.4.7
certifi==2020.6.20
cffi==1.14.3
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.2.1
cryptography==3.1.1
cssutils==1.0.2
daphne==2.5.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2
django-allauth==0.42.0
django-celery-beat==2.0.0
django-celery-results==1.2.1
django-channels-presence==1.0.0
django-cors-headers==3.5.0
django-inline-static==0.1.0
django-inlinecss==0.3.0
django-money==1.1
django-oauth-toolkit==1.3.2
django-picklefield==3.0.1
django-q==1.3.3
django-rest-auth==0.9.5
django-rest-swagger==2.2.0
django-safedelete==0.5.6
django-softdelete==0.9.1
django-timezone-field==4.0
djangorestframework==3.11.1
djangorestframework-camel-case==1.2.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
dnspython==1.16.0
drf-yasg==1.17.1
eventlet==0.28.0
future==0.18.2
google-auth==1.22.0
greenlet==0.4.17
gunicorn==20.0.4
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
idna-ssl==1.1.0
importlib-metadata==2.0.0
incremental==17.5.0
inflection==0.5.1
install==1.3.4
itypes==1.2.0
Jinja2==2.11.2
jinxed==1.0.1
kombu==4.6.11
MarkupSafe==1.1.1
monotonic==1.5
msgpack==0.6.2
multidict==4.7.6
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.4
Pillow==8.0.1
prompt-toolkit==3.0.8
psycopg2==2.8.6
py-moneyed==0.8.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pynliner==0.8.0
pyOpenSSL==19.1.0
pyotp==2.4.1
pyparsing==2.4.7
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
service-identity==18.1.0
simplejson==3.17.2
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.3
ua-parser==0.10.0
uritemplate==3.0.1
urllib3==1.25.10
user-agents==2.2.0
vine==1.3.0
wcwidth==0.2.5
whitenoise==5.2.0
yarl==1.6.0
zipp==3.3.0
zope.interface==5.1.2
